--- OstreCWEB/OstreCWEB.Services/StoryReader/IStoryReaderServices.cs
@@ -1,6 +1,0 @@
-﻿namespace OstreCWEB.Services.StoryReader
-{
-    internal interface IStoryReaderServices
-    {
-    }
-}
