--- OstreCWEB/OstreCWEB.Services/StoryReader/StoryReaderServices.cs
@@ -1,6 +1,0 @@
-﻿namespace OstreCWEB.Services.StoryReader
-{
-    internal class StoryReaderServices
-    {
-    }
-}
