+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/CreatNewParagraphView.cs
@@ -1,7 +1,7 @@
 ﻿using OstreCWEB.DomainModels.StoryModels.Enums;
 using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.Services.StoryBuilder.ModelsDto
+namespace OstreCWEB.Services.StoryService.ModelsDto
 {
     public class CreatNewParagraphView
     {
