+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/EnemyInParagraphService.cs
@@ -1,16 +1,13 @@
-﻿namespace OstreCWEB.Services.StoryBuilder.ModelsDto
+﻿namespace OstreCWEB.Services.StoryService.Models
 {
-    public class EnemyInParagraphView
+    public class EnemyInParagraphService
     {
         public int Id { get; set; }
-
-        public string EnemyName { get; set; }
         public int AmountOfEnemy { get; set; }
+        public string EnemyName { get; set; }
 
         public int ParagraphId { get; set; }
         public int FightPropId { get; set; }
         public int EnemyId { get; set; }
-
-        public Dictionary<int, string> Enemies { get; set; }
     }
 }
