+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/EnemyInParagraphView.cs
@@ -1,13 +1,16 @@
-﻿namespace OstreCWEB.Services.StoryBuilder.Models
+﻿namespace OstreCWEB.Services.StoryService.ModelsDto
 {
-    public class EnemyInParagraphService
+    public class EnemyInParagraphView
     {
         public int Id { get; set; }
-        public int AmountOfEnemy { get; set; }
+
         public string EnemyName { get; set; }
+        public int AmountOfEnemy { get; set; }
 
         public int ParagraphId { get; set; }
         public int FightPropId { get; set; }
         public int EnemyId { get; set; }
+
+        public Dictionary<int, string> Enemies { get; set; }
     }
 }
