+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/ParagraphDetailsView.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.Services.StoryBuilder.ModelsDto
+namespace OstreCWEB.Services.StoryService.ModelsDto
 {
     public class ParagraphDetailsView
     {
