+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/ParagraphElementView.cs
@@ -1,9 +1,9 @@
 ﻿using OstreCWEB.DomainModels.StoryModels.Enums;
 using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.Services.StoryBuilder.ModelsDto
+namespace OstreCWEB.Services.StoryService.ModelsDto
 {
-    public class ParagraphWithCoiceView
+    public class ParagraphElementView
     {
         public int Id { get; set; }
 
@@ -16,7 +16,6 @@ namespace OstreCWEB.Services.StoryBuilder.ModelsDto
         [Display(Name = "Amount Of Choices")]
         public int AmountOfChoices { get; set; }
 
-        public int ChoiceId { get; set; }
-        public string DescriptionOfChoice { get; set; }
+        public int StoryId { get; set; }
     }
 }
