+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/ParagraphWithCoice.cs
@@ -1,6 +1,6 @@
 ﻿using OstreCWEB.DomainModels.StoryModels.Enums;
 
-namespace OstreCWEB.Services.StoryBuilder.Models
+namespace OstreCWEB.Services.StoryService.Models
 {
     public class ParagraphWithCoice
     {
