+++ OstreCWEB/OstreCWEB.Services/StoryService/ModelsDto/StoryParagraphsView.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.Services.StoryBuilder.ModelsDto
+namespace OstreCWEB.Services.StoryService.ModelsDto
 {
     public class StoryParagraphsView
     {
