+++ OstreCWEB/OstreCWEB/Mapping/GameProfile.cs
@@ -4,7 +4,7 @@ using OstreCWEB.DomainModels.CharacterModels;
 using OstreCWEB.DomainModels.ManyToMany;
 using OstreCWEB.DomainModels.StoryModels;
 using OstreCWEB.DomainModels.StoryModels.Properties;
-using OstreCWEB.Services.StoryBuilder.ModelsDto;
+using OstreCWEB.Services.StoryService.ModelsDto;
 using OstreCWEB.ViewModel.Game;
 using OstreCWEB.ViewModel.Identity;
 using OstreCWEB.ViewModel.StoryReader;
