+++ OstreCWEB/OstreCWEB/ViewModel/Identity/UserView.cs
@@ -1,5 +1,5 @@
 ﻿using OstreCWEB.ViewModel.Characters;
-using OstreCWEB.Services.StoryBuilder.ModelsDto;
+using OstreCWEB.Services.StoryService.ModelsDto;
 using System.ComponentModel;
 
 namespace OstreCWEB.ViewModel.Identity
