+++ OstreCWEB/OstreCWEB/Views/Game/_StoryRow.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<OstreCWEB.Services.StoryBuilder.ModelsDto.StoryView>
+﻿@model IEnumerable<OstreCWEB.Services.StoryService.ModelsDto.StoryView>
 @{
     ViewData["Title"] = "_PlayableCharacterTable";
 }
