+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/AddEnemyInParagraph.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.EnemyInParagraphView
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.EnemyInParagraphView
 
 @{
     ViewData["Title"] = "Add Enemy to Fight";
