+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ChangeParagraph.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.ChoiceCreatorView
+@model OstreCWEB.Services.StoryService.ModelsDto.ChoiceCreatorView
 
 @{
     ViewData["Title"] = $"Edit Choice";
