+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ChangeSecondParagraph.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.StoryParagraphsView 
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.StoryParagraphsView 
 @using OstreCWEB.DomainModels.StoryModels.Enums
 @{
     ViewData["Title"] = $"Story " + @Model.Name;
