+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ChoiceDetails.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.ChoiceDetailsView
+@model OstreCWEB.Services.StoryService.ModelsDto.ChoiceDetailsView
 
 @{
     ViewData["Title"] = $"Story " + @Model.NameOfStory;
