+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ChooseSecondParagraph.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.StoryParagraphsView
+@model OstreCWEB.Services.StoryService.ModelsDto.StoryParagraphsView
 
 @{
     ViewData["Title"] = $"Story " + @Model.Name;
