+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatChoice.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums 
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.ChoiceCreatorView
+@model OstreCWEB.Services.StoryService.ModelsDto.ChoiceCreatorView
 
 @{
     ViewData["Title"] = $"Choice Creator";
