+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatNewParagraph.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.CreatNewParagraphView
+@model OstreCWEB.Services.StoryService.ModelsDto.CreatNewParagraphView
 
 @{
     ViewData["Title"] = "Create Paragraph";
