+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatParagraphFight.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.CreatParagraphFightView
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.CreatParagraphFightView
 
 @{
     ViewData["Title"] = "Create Paragraph Fight";
