+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatParagraphTest.cshtml
@@ -1,6 +1,6 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
 @using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.CreatParagraphTestView
+@model OstreCWEB.Services.StoryService.ModelsDto.CreatParagraphTestView
 
 @{
     ViewData["Title"] = "Create Paragraph Test";
