+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreateStory.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.StoryView
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.StoryView
 
 @{
     ViewData["Title"] = "Create story";
