+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/DeleteParagraph.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.ParagraphElementView
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.ParagraphElementView
 
 @{
     ViewData["Title"] = "Delete Paragraph";
