+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/EditChoice.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.ChoiceCreatorView
+@model OstreCWEB.Services.StoryService.ModelsDto.ChoiceCreatorView
 
 @{
     ViewData["Title"] = $"Edit Choice";
