+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/EditParagraph.cshtml
@@ -1,6 +1,6 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
 @using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.Services.StoryBuilder.ModelsDto.EditParagraphView
+@model OstreCWEB.Services.StoryService.ModelsDto.EditParagraphView
 
 @{
     ViewData["Title"] = "Paragraph editor";
