+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/EditStory.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.StoryView
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.StoryView
 
 @{
     ViewData["Title"] = "Edit Story";
