+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<OstreCWEB.Services.StoryBuilder.ModelsDto.StoryView>
+﻿@model IEnumerable<OstreCWEB.Services.StoryService.ModelsDto.StoryView>
 
 @{
     ViewData["Title"] = "StoryBuilder";
