+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ParagraphDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.Services.StoryBuilder.ModelsDto.ParagraphDetailsView
+﻿@model OstreCWEB.Services.StoryService.ModelsDto.ParagraphDetailsView
 
 @{
     ViewData["Title"] = $"Story " + @Model.NameOfStory;
